requirements.txt
fastapi==0.115.6
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.12
supabase==2.10.0
httpx==0.27.2
python-dotenv==1.0.1
